requirements.txt
openai
openai-whisper
requests
Pillow
moviepy
google-generativeai
google-cloud-aiplatform
google-cloud-texttospeech
spleeter
